+++ crates/swc_transformer_import_glob/tests/mod.rs
@@ -1,6 +1,11 @@
-use std::{collections::HashMap, sync::Arc};
+use std::sync::Arc;
 
-use farmfe_core::{swc_common::SourceMap, swc_ecma_ast::EsVersion, swc_ecma_parser::Syntax};
+use farmfe_core::{
+  config::{AliasItem, StringOrRegex},
+  swc_common::SourceMap,
+  swc_ecma_ast::EsVersion,
+  swc_ecma_parser::Syntax,
+};
 use farmfe_swc_transformer_import_glob::transform_import_meta_glob;
 use farmfe_testing_helpers::fixture;
 use farmfe_toolkit::script::{codegen_module, parse_module, ParseScriptModuleResult};
@@ -29,16 +34,16 @@ fn test_import_meta_glob() {
       &mut ast,
       root.to_string(),
       dir.to_string(),
-      &HashMap::from([(
-        "@".to_string(),
-        file
+      &vec![AliasItem::Complex {
+        find: StringOrRegex::String("@".to_string()),
+        replacement: file
           .parent()
           .unwrap()
           .to_path_buf()
           .join("dir")
           .to_string_lossy()
           .to_string(),
-      )]),
+      }],
     )
     .unwrap();