+++ js-plugins/sass/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @farmfe/js-plugin-sass
 
+## 3.0.0-nightly-20241022041556
+
+### Patch Changes
+
+- Updated dependencies [24347c9]
+  - @farmfe/core@2.0.0-nightly-20241022041556
+
+## 3.0.0-nightly-20241022014521
+
+### Patch Changes
+
+- Updated dependencies [a82a5b2]
+  - @farmfe/core@2.0.0-nightly-20241022014521
+
 ## 2.9.0
 
 ### Patch Changes